--- Hypercube.Shared/EventBus/Events/Unit.cs
@@ -1,11 +1,0 @@
-﻿using System.Runtime.InteropServices;
-
-namespace Hypercube.Shared.EventBus.Events;
-
-public readonly struct Unit;
-
-[StructLayout(LayoutKind.Sequential)]
-public sealed class UnitBox
-{
-    public Unit Value;
-}
