+++ Hypercube.Shared/Utilities/Units/Unit.cs
@@ -1,0 +1,8 @@
+﻿using System.Runtime.CompilerServices;
+
+namespace Hypercube.Shared.Utilities.Units;
+
+/// <summary>
+/// Whenever you see this struct it some other object, it should be resolved using Unsafe.As()
+/// </summary>
+public readonly struct Unit;
